+++ lib/ui/detail/view/detail_page.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:gap/gap.dart';
-import 'package:sticky_headers/sticky_headers.dart';
-import 'package:weather_app_flutter_mvvm/extension/int_extension.dart';
-import 'package:weather_app_flutter_mvvm/model/response/weathre_response_model.dart';
 import 'package:weather_app_flutter_mvvm/ui/detail/component/pop_chart.dart';
+import 'package:weather_app_flutter_mvvm/ui/detail/component/weather_list.dart';
 import 'package:weather_app_flutter_mvvm/ui/detail/view_model/detail_view_model.dart';
 
 class DetailPage extends ConsumerStatefulWidget {
@@ -60,47 +57,7 @@ class _DetailPageState extends ConsumerState<DetailPage> {
                 ),
               ),
               Expanded(
-                child: ListView.builder(
-                  itemCount: data.threeHoursWeather.length,
-                  itemBuilder: (context, index) {
-                    String date = data.threeHoursWeather.keys.elementAt(index);
-                    List<WeatherData>? threeHourWethers =
-                        data.threeHoursWeather[date];
-                    if (threeHourWethers == null) {
-                      return const Text('データがありません');
-                    }
-                    return StickyHeader(
-                      header: Container(
-                        height: 30.0,
-                        color: const Color.fromARGB(249, 244, 244, 244),
-                        padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                        alignment: Alignment.centerLeft,
-                        child: Text(
-                          date,
-                        ),
-                      ),
-                      content: Column(
-                        children: threeHourWethers
-                            .map(
-                              (event) => Padding(
-                                padding:
-                                    const EdgeInsets.symmetric(vertical: 10),
-                                child: _WeatherListCell(
-                                  maxTemperature:
-                                      event.main.tempMax.toStringAsFixed(1),
-                                  minTemperature:
-                                      event.main.tempMin.toStringAsFixed(1),
-                                  humidityLevel: event.main.humidity.toString(),
-                                  imageName: "01",
-                                  time: event.dt.toStringHHMMFromEpoch(),
-                                ),
-                              ),
-                            )
-                            .toList(),
-                      ),
-                    );
-                  },
-                ),
+                child: WeatherList(weathreData: data.threeHoursWeather),
               ),
             ],
           );
@@ -111,68 +68,6 @@ class _DetailPageState extends ConsumerState<DetailPage> {
             child: CircularProgressIndicator(),
           );
         }),
-      ),
-    );
-  }
-}
-
-class _WeatherListCell extends StatelessWidget {
-  final String time;
-  final String imageName;
-  final String maxTemperature;
-  final String minTemperature;
-  final String humidityLevel;
-  const _WeatherListCell({
-    required this.maxTemperature,
-    required this.minTemperature,
-    required this.humidityLevel,
-    required this.imageName,
-    required this.time,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20),
-      child: Row(
-        children: [
-          Text(time),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 10),
-            child: Image.asset(
-              'assets/images/splash_logo.png',
-              width: 50,
-              height: 50,
-            ),
-          ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                '最高気温:$maxTemperature ℃',
-                style: const TextStyle(
-                  color: Colors.red,
-                ),
-              ),
-              const SizedBox(
-                height: 5,
-              ),
-              Text(
-                '最低気温:$minTemperature ℃',
-                style: const TextStyle(
-                  color: Colors.blue,
-                ),
-              ),
-              const Gap(5),
-              Text(
-                '湿度:$humidityLevel %',
-                style: const TextStyle(
-                  color: Colors.green,
-                ),
-              ),
-            ],
-          )
-        ],
       ),
     );
   }
